pyproject.toml
[project]
name = "clepsy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "asyncclick>=8.1.7.2",
    "jsonref>=1.1.0",
    "levenshtein>=0.26.1",
    "pillow>=11.0.0",
    "pydantic-settings>=2.6.0",
    "tenacity>=9.0.0",
    "fastapi>=0.111.1",
    "htpy>=24.10.1",
    "aiocache>=0.12.3",
    "unidecode>=1.3.8",
    "openai>=1.61.1",
    "msgpack>=1.1.0",
    "google-genai>=1.2.0",
    "aiomcache>=0.8.2",
    "ujson>=5.10.0",
    "tzlocal>=5.3",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.2",
    "python-dateutil>=2.9.0",
    "baml-py==0.212.0",
    "loguru>=0.7.3",
    "argon2-cffi>=25.1.0",
    "sqlalchemy>=2.0.43",
    "apscheduler[sqlalchemy]>=4.0.0.0a6",
    "paddleocr>=3.2.0",
    "paddlepaddle>=3.2.0",
    "torch>=2.0.0",
    "gliner>=0.2.22",
]

[dependency-groups]
dev = [
    "openapi-generator-cli>=7.13.0",
    "pytailwindcss>=0.2.0",
    "ruff>=0.8.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "pytest-cov>=6.0.0",
    "pytest-sugar>=1.0.0",
    "pytest-emoji>=0.2.0",
    "rich>=13.0.0",
]

[tool.uv]
python-preference = "managed"




[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "paddlepaddle"
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu/"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
paddlepaddle = { index = "paddlepaddle" }



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]

select = ["E4", "E7", "E9", "F", "I"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
known-first-party = ["clepsy", "baml_client"]
combine-as-imports = true
force-sort-within-sections = true
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"


docstring-code-format = false


docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = "."
addopts = "--color=yes --tb=short -v --strict-markers --strict-config --emoji"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_auto_indent = true
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore:.*urllib3.*:DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "llm: marks tests that rely on live LLM or BAML model calls",
    "pii: marks tests that require PII detection model (GLiNER)",
]


[tool.pylint."MESSAGES CONTROL"]
disable = [
    "logging-fstring-interpolation",
  "missing-function-docstring",
  "missing-class-docstring",
  "wrong-import-order",
  "missing-module-docstring",
  "dangerous-default-value"
]
